requirements.txt
streamlit==1.44.1

langchain==0.3.23
langchain-community==0.3.21
langchain-groq==0.3.2

faiss-cpu==1.8.0

pypdf==5.4.0

sentence-transformers==3.4.1
transformers==4.48.3
tokenizers==0.21.0

torch==2.11.0+cpu

python-dotenv==1.1.0
deep-translator==1.11.4

numpy==1.26.4
pandas==2.3.3